requirements.txt
flask
flask_httpauth
flask_socketio
python-dotenv
websocket
flask-limiter